refactor: 파일당 하나의 컴포넌트 원칙 적용
- ChatMessageList 서브컴포넌트 분리 (AIAvatar, UserMessage, AIMessage 등)
- ExperienceCard(chat) 서브컴포넌트 분리 (CategoryTag, HashtagBadge)
- HomeClient 서브컴포넌트 분리 (ExperienceCard, SectionHeader, DesignTokensTest)
- ProjectList 서브컴포넌트 분리 (ProjectRow, ProjectListSkeleton)
- page.tsx에서 ProjectListServer 분리
- CLAUDE.md에 컴포넌트 분리 규칙 문서화
